feat: Allow users to choose clone location
Introduces functionality for users to select a directory where a repository will be cloned. This is achieved by adding a new IPC handler for showing the directory picker and integrating it into the cloning process. The UI has been updated to include a new button on the repository card for this action.