feat: Introduce command selection for launch tasks
This change refactors the `LaunchConfig` to support selecting executables or running direct commands. It also introduces a new modal for selecting executables and migrates existing `InstallDeps` and `RunTests` task types to the generic `RunCommand` type, using the detected package manager. Global settings for default package manager are removed as they are now inferred.